Cast to double to solve bug computing number of iterations on large problems.

+++ trunk/fftw++/tconv.cc
@@ -100,7 +100,7 @@ int main(int argc, char* argv[])
   cout << "m=" << m << endl;
   
   if(N == 0) {
-    N=N0/n;
+    N=(int) ((double)N0/(double)n);
     if(N < 10) N=10;
   }
   cout << "N=" << N << endl;
